int ret instead of long now the functions are int too

+++ tpe_template.c
@@ -172,7 +172,7 @@ int tpe_allow(const char *name) {
 int tpe_do_execve(char __user *name, char __user * __user *argv,
 		char __user * __user *envp, struct pt_regs *regs) {
 
-	long ret;
+	int ret;
 
 	ret = tpe_allow(name);
 
@@ -193,7 +193,7 @@ int tpe_do_execve(char __user *name, char __user * __user *argv,
 int tpe_compat_do_execve(char __user *name, char __user * __user *argv,
 		char __user * __user *envp, struct pt_regs *regs) {
 
-	long ret;
+	int ret;
 
 	ret = tpe_allow(name);
